+++ 21YearsOfGasPricesDescriptive/21YearsOfGasPrices/NewLine.h
@@ -7,13 +7,13 @@ public:
 
 	// declare the NewLine member variables
 	int index = 0, month = 0, day = 0, year = 0, size = 1065;
-	float price = 0;
+	double price = 0;
 
 	// declare the default constructor
 	NewLine() {}
 
 	// set the member variables to the input values
-	void setValues(int month_, int day_, int year_, float price_) {
+	void setValues(int month_, int day_, int year_, double price_) {
 		month = month_, day = day_, year = year_, price = price_;
 	}
